pyproject.toml
[tool.poetry]
name = "lacss"
version = "0.4.4"
description = "Cell segmentation and tracking"
authors = ["Ji Yu <jyu@uchc.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
flax = "^0.7.2"
tqdm = "^4.65.0"
imageio = "^2.9.0"
scikit-image = ">=0.19.0"
typer = ">=0.4.0"
imagecodecs = "^2023.3.16"
pycocotools = "^2.0.6"
chex = "^0.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
jupyterlab = "^3.6.1"
jax = {extras = ["cuda11-pip"], version = "^0.4.14"}
tensorflow-cpu = "^2.13.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.2.3"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
mkdocs-gen-files = "^0.5.0"

[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
default = false
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
